refactor: extract openai stream wrappers
